+++ packages/auth/src/providers/cognito/credentialsProvider/credentialsProvider.ts
@@ -15,6 +15,7 @@ import {
 } from '@aws-amplify/core/internals/utils';
 
 import { AuthError } from '../../../errors/AuthError';
+import { assertServiceError } from '../../../errors/utils/assertServiceError';
 import { getRegionFromIdentityPoolId } from '../../../foundation/parsers';
 import { assertIdTokenInAuthTokens } from '../utils/types';
 
@@ -116,19 +117,27 @@ export class CognitoAWSCredentialsAndIdentityIdProvider
 		// save credentials in-memory
 		// No logins params should be passed for guest creds:
 		// https://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_GetCredentialsForIdentity.html
-		const clientResult = await getCredentialsForIdentity(
-			{ region },
-			{
-				IdentityId: identityId,
-			},
-		);
+		let clientResult:
+			| Awaited<ReturnType<typeof getCredentialsForIdentity>>
+			| undefined;
+		try {
+			clientResult = await getCredentialsForIdentity(
+				{ region },
+				{
+					IdentityId: identityId,
+				},
+			);
+		} catch (e) {
+			assertServiceError(e);
+			throw new AuthError(e);
+		}
 
 		if (
-			clientResult.Credentials &&
-			clientResult.Credentials.AccessKeyId &&
-			clientResult.Credentials.SecretKey
+			clientResult?.Credentials?.AccessKeyId &&
+			clientResult?.Credentials?.SecretKey
 		) {
 			this._nextCredentialsRefresh = new Date().getTime() + CREDENTIALS_TTL;
+
 			const res: CredentialsAndIdentityId = {
 				credentials: {
 					accessKeyId: clientResult.Credentials.AccessKeyId,
@@ -138,11 +147,10 @@ export class CognitoAWSCredentialsAndIdentityIdProvider
 				},
 				identityId,
 			};
-			const identityIdRes = clientResult.IdentityId;
-			if (identityIdRes) {
-				res.identityId = identityIdRes;
+			if (clientResult.IdentityId) {
+				res.identityId = clientResult.IdentityId;
 				this._identityIdStore.storeIdentityId({
-					id: identityIdRes,
+					id: clientResult.IdentityId,
 					type: 'guest',
 				});
 			}
@@ -186,19 +194,28 @@ export class CognitoAWSCredentialsAndIdentityIdProvider
 
 		const region = getRegionFromIdentityPoolId(authConfig.identityPoolId);
 
-		const clientResult = await getCredentialsForIdentity(
-			{ region },
-			{
-				IdentityId: identityId,
-				Logins: logins,
-			},
-		);
+		let clientResult:
+			| Awaited<ReturnType<typeof getCredentialsForIdentity>>
+			| undefined;
+		try {
+			clientResult = await getCredentialsForIdentity(
+				{ region },
+				{
+					IdentityId: identityId,
+					Logins: logins,
+				},
+			);
+		} catch (e) {
+			assertServiceError(e);
+			throw new AuthError(e);
+		}
 
 		if (
-			clientResult.Credentials &&
-			clientResult.Credentials.AccessKeyId &&
-			clientResult.Credentials.SecretKey
+			clientResult?.Credentials?.AccessKeyId &&
+			clientResult?.Credentials?.SecretKey
 		) {
+			this._nextCredentialsRefresh = new Date().getTime() + CREDENTIALS_TTL;
+
 			const res: CredentialsAndIdentityId = {
 				credentials: {
 					accessKeyId: clientResult.Credentials.AccessKeyId,
@@ -208,22 +225,22 @@ export class CognitoAWSCredentialsAndIdentityIdProvider
 				},
 				identityId,
 			};
+
+			if (clientResult.IdentityId) {
+				res.identityId = clientResult.IdentityId;
+				// note: the following call removes guest identityId from the persistent store (localStorage)
+				this._identityIdStore.storeIdentityId({
+					id: clientResult.IdentityId,
+					type: 'primary',
+				});
+			}
+
 			// Store the credentials in-memory along with the expiration
 			this._credentialsAndIdentityId = {
 				...res,
 				isAuthenticatedCreds: true,
 				associatedIdToken: authTokens.idToken?.toString(),
 			};
-			this._nextCredentialsRefresh = new Date().getTime() + CREDENTIALS_TTL;
-
-			const identityIdRes = clientResult.IdentityId;
-			if (identityIdRes) {
-				res.identityId = identityIdRes;
-				this._identityIdStore.storeIdentityId({
-					id: identityIdRes,
-					type: 'primary',
-				});
-			}
 
 			return res;
 		} else {