+++ packages/storage/src/providers/s3/utils/client/s3data/headObject.ts
@@ -70,8 +70,7 @@ const headObjectDeserializer = async (
 ): Promise<HeadObjectOutput> => {
 	if (response.statusCode >= 300) {
 		// error is always set when statusCode >= 300
-		const error = (await parseXmlError(response)) as Error;
-		throw buildStorageServiceError(error, response.statusCode);
+		throw buildStorageServiceError((await parseXmlError(response))!);
 	} else {
 		const contents = {
 			...map(response.headers, {