+++ packages/geo/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.80](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.79...@aws-amplify/geo@3.0.80) (2025-04-28)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.79](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.78...@aws-amplify/geo@3.0.79) (2025-04-21)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.78](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.77...@aws-amplify/geo@3.0.78) (2025-04-09)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.77](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.76...@aws-amplify/geo@3.0.77) (2025-03-28)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.76](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.75...@aws-amplify/geo@3.0.76) (2025-03-25)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.75](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.74...@aws-amplify/geo@3.0.75) (2025-03-21)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.74](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.73...@aws-amplify/geo@3.0.74) (2025-03-14)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.73](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.72...@aws-amplify/geo@3.0.73) (2025-03-07)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+## [3.0.72](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.71...@aws-amplify/geo@3.0.72) (2025-03-05)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
 ## [3.0.71](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.70...@aws-amplify/geo@3.0.71) (2025-02-27)
 
 **Note:** Version bump only for package @aws-amplify/geo